futures: {read,write}: add seekable flag
When calling `read` or `write` via io_uring, we must specify whether the
underlying descriptors offset needs to be incremented or not by setting
the `offset` flag to `-1`/`u32::MAX. Note that for non-seekable fds this
value should be set to `0`. Therefore, add a new flag that creators of
the Async{Read,Write} structs can set to indicate wether the underlying
file is seekable or not.